Issue #259 refactor packages
