+++ api/src/main/java/jakarta/data/Direction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Contributors to the Eclipse Foundation
+ * Copyright (c) 2022,2023 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
  *  SPDX-License-Identifier: Apache-2.0
  */
 
-package jakarta.data.repository;
+package jakarta.data;
 
 /**
  * Represents the direction of the ordering.
