+++ api/src/main/java/jakarta/data/page/KeysetCursor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Contributors to the Eclipse Foundation
+ * Copyright (c) 2022,2023 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,13 @@
  *
  *  SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.repository;
+package jakarta.data.page;
 
 import java.util.Arrays;
 
-    /**
-     * Built-in implementation of Cursor for keyset pagination.
-     */
+/**
+ * Built-in implementation of Cursor for keyset pagination.
+ */
 class KeysetCursor implements Pageable.Cursor {
     /**
      * Keyset values.
