+++ api/src/main/java/module-info.java
@@ -16,14 +16,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import jakarta.data.Limit;
+import jakarta.data.Sort;
+import jakarta.data.page.Pageable;
 import jakarta.data.repository.CrudRepository;
 import jakarta.data.repository.DataRepository;
-import jakarta.data.repository.Limit;
 import jakarta.data.repository.OrderBy;
-import jakarta.data.repository.Pageable;
 import jakarta.data.repository.Query;
 import jakarta.data.repository.Repository;
-import jakarta.data.repository.Sort;
 
 /**
  * <p>Jakarta Data standardizes a programming model where data is represented by simple Java classes
@@ -654,6 +654,8 @@ import jakarta.data.repository.Sort;
 //       under: "Wildcard characters for patterns are determined by the data access provider"
 // TODO Ensure we have all required supported return types listed.
 module jakarta.data {
+    exports jakarta.data;
+    exports jakarta.data.page;
     exports jakarta.data.repository;
     exports jakarta.data.exceptions;
     opens jakarta.data.repository;
