+++ tck/src/main/java/ee/jakarta/tck/data/framework/read/only/AsciiCharacters.java
@@ -20,12 +20,12 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Stream;
 
+import jakarta.data.Sort;
+import jakarta.data.Streamable;
+import jakarta.data.page.Page;
+import jakarta.data.page.Pageable;
 import jakarta.data.repository.DataRepository;
-import jakarta.data.repository.Page;
-import jakarta.data.repository.Pageable;
 import jakarta.data.repository.Repository;
-import jakarta.data.repository.Sort;
-import jakarta.data.repository.Streamable;
 
 /**
  * This is a read only repository that represents the set of AsciiCharacters from 0-256.
